Use xattrs to avoid redundant MD5 sum generation
Adds an extended attribute to the NetCDF file (not a NetCDF attribute)
to avoid needing to regenerate MD5 sums for files each and every time.
This commit addresses a long standing performance issue where *all* the
archived files would have MD5 sums calculated upon each archival run,
even if the contents of the file were not changing.  In the case of
files being reuploaded, this will be reflected in the hard link content.